Creating new FooterStyled document for shared footer design
