+++ source/frontend/src/components/layout/Footer/EmptyFooter.tsx
@@ -1,6 +1,6 @@
-import styled from 'styled-components';
-
 import ApiVersionInfo from '@/components/common/ApiVersionInfo';
+
+import FooterStyled from './FooterStyled';
 
 /**
  * Display an "empty" header bar with limited functionality as a placeholder
@@ -8,77 +8,39 @@ import ApiVersionInfo from '@/components/common/ApiVersionInfo';
 const EmptyFooter = () => {
   return (
     <FooterStyled>
-      <a
-        target="_blank"
-        rel="noopener noreferrer"
-        href="http://www.gov.bc.ca/gov/content/home/disclaimer"
-      >
-        Disclaimer
-      </a>
-      <a
-        target="_blank"
-        rel="noopener noreferrer"
-        href="http://www.gov.bc.ca/gov/content/home/privacy"
-      >
-        Privacy
-      </a>
-      <a
-        target="_blank"
-        rel="noopener noreferrer"
-        href="http://www.gov.bc.ca/gov/content/home/accessible-government"
-      >
-        Accessibility
-      </a>
-      <a
-        target="_blank"
-        rel="noopener noreferrer"
-        href="http://www.gov.bc.ca/gov/content/home/copyright"
-      >
-        Copyright
-      </a>
-      <ApiVersionInfo />
+      <div className="main-footer">
+        <a
+          target="_blank"
+          rel="noopener noreferrer"
+          href="http://www.gov.bc.ca/gov/content/home/disclaimer"
+        >
+          Disclaimer
+        </a>
+        <a
+          target="_blank"
+          rel="noopener noreferrer"
+          href="http://www.gov.bc.ca/gov/content/home/privacy"
+        >
+          Privacy
+        </a>
+        <a
+          target="_blank"
+          rel="noopener noreferrer"
+          href="http://www.gov.bc.ca/gov/content/home/accessible-government"
+        >
+          Accessibility
+        </a>
+        <a
+          target="_blank"
+          rel="noopener noreferrer"
+          href="http://www.gov.bc.ca/gov/content/home/copyright"
+        >
+          Copyright
+        </a>
+        <ApiVersionInfo />
+      </div>
     </FooterStyled>
   );
 };
-
-/**
- * Styled footer component.
- */
-const FooterStyled = styled('div')`
-  width: 95%;
-  padding: 1rem 20%;
-  display: flex;
-  flex-direction: row;
-  flex-wrap: wrap;
-  justify-content: flex-start;
-  align-items: stretch;
-  align-content: flex-start;
-  font-size: 1.2rem;
-
-  a {
-    color: #ffffff;
-    display: inline;
-    &:link,
-    &:visited {
-      font-size: 1.2rem;
-      font-family: 'BCSans', Fallback, sans-serif;
-      padding: 0 1rem;
-      border-right: 0.1rem solid #666;
-    }
-    &:hover,
-    &:focus {
-      text-decoration: underline;
-    }
-  }
-
-  div.version {
-    color: #ffffff;
-    font-size: 1.2rem;
-    font-family: 'BCSans', Fallback, sans-serif;
-    padding: 0 0 0 1rem;
-    flex-grow: 1;
-    text-align: right;
-  }
-`;
 
 export default EmptyFooter;
